refactor: アプリケーション層をsrc/apps/ディレクトリに集約
- cli, tui, mcpディレクトリをsrc/apps/下に移動
- アプリケーション層とドメイン層の境界を明確化
- すべてのimportパスを新しい構造に合わせて更新
- package.jsonのmigrateスクリプトパスを更新
- CLAUDE.mdのアーキテクチャ説明を更新

この変更により、プロジェクトの構造がより整理され、
どのディレクトリがアプリケーションのエントリーポイントなのか
一目で分かるようになりました。
