+++ src/apps/cli/commands/index.ts
