+++ src/apps/cli/commands/tui.tsx
