+++ src/apps/mcp/resources/index.ts
