+++ src/apps/mcp/server.ts
@@ -3,13 +3,13 @@ import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js'
 import { readFileSync } from 'node:fs';
 import { fileURLToPath } from 'node:url';
 import { dirname, join } from 'node:path';
-import { ArticleModel } from '../models/article.js';
-import { FeedModel } from '../models/feed.js';
+import { ArticleModel } from '../../models/article.js';
+import { FeedModel } from '../../models/feed.js';
 import { registerArticleResources } from './resources/articles.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
-const packageJson = JSON.parse(readFileSync(join(__dirname, '../../package.json'), 'utf-8')) as {
+const packageJson = JSON.parse(readFileSync(join(__dirname, '../../../package.json'), 'utf-8')) as {
   name: string;
   version: string;
 };
