+++ src/apps/mcp/types.ts
