+++ src/apps/tui/App.autoread.test.tsx
