+++ src/apps/tui/App.tsx
@@ -2,9 +2,9 @@ import { Box, Text, useApp, useStdout } from 'ink';
 import { useCallback, useEffect, useState, useMemo } from 'react';
 import { spawn } from 'child_process';
 import type { Article, Feed } from '@/types';
-import { FeedService } from '../services/feed-service.js';
-import { FeedModel } from '../models/feed.js';
-import { ArticleModel } from '../models/article.js';
+import { FeedService } from '../../services/feed-service.js';
+import { FeedModel } from '../../models/feed.js';
+import { ArticleModel } from '../../models/article.js';
 import { createDatabaseManager } from '../cli/utils/database.js';
 import { ArticleList } from './components/ArticleList.js';
 import { FeedList } from './components/FeedList.js';
