+++ src/apps/tui/hooks/useKeyboardNavigation.test.tsx
