+++ src/apps/tui/hooks/useKeyboardNavigation.ts
