+++ src/apps/tui/utils/html.ts
